Update project for package renaming
- Changed package name from "rtui-app" to "rtui2" in pyproject.toml and README.
- Updated authorship information and copyright year in LICENSE and NOTICE files.
- Introduced new files and modules for ROS 2 functionality, including entity inspection and event handling.
- Removed legacy ROS 1 interface files and tests.
- Enhanced README with installation instructions and project details.
