+++ pyproject.toml
@@ -1,30 +1,25 @@
 [tool.poetry]
-name = "rtui-app"
-version = "0.2.1"
-description = "A TUI tool for ROS1/2"
-authors = ["Yuma Hiramatsu <yuma.hiramatsu@gmail.com>"]
-repository = "https://github.com/eduidl/rtui"
-license = "Apache License 2.0"
+name = "rtui2"
+version = "0.1.0"
+description = "A TUI tool for ROS 2 Topic/Node debugging"
+authors = ["Daisuke Sato <daisuke.sato@whill.inc>"]
+repository = "https://github.com/whill-labs/rtui2"
+license = "Apache License, Version 2.0"
 readme = "README.md"
 
 [tool.poetry.scripts]
-rtui = "rtui_app.cli:main"
+rtui = "rtui2.cli:main"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.10"
 click = "^8.1.3"
 textual = "^0.40"
 typing-extensions = "^4.3.0"
 
-# ROS1/2
 netifaces = "^0.11.0"
-# ROS1
-PyYAML = "^6.0"
-rospkg = "^1.4.0"
-# ROS2
 numpy = "^1.23.0"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 black = "^23"
 flake8 = { version = "^6", python = "^3.10" }
 flake8-pyproject = "^1"
@@ -35,9 +30,9 @@ taskipy = "^1"
 [tool.taskipy.tasks]
 check_all = "task check && task check_type"
 check = "task check_lint && task check_format"
-check_lint = "flake8 rtui_app"
+check_lint = "flake8 rtui2"
 check_format = "black . --check && isort . --check --diff"
-check_type = "mypy rtui_app"
+check_type = "mypy rtui2"
 format = "black . && isort ."
 test = "python -m unittest tests -v"
