+++ rtui2/app/inspect.py
