+++ rtui2/ros/entity.py
