+++ rtui2/widgets/info_panel.py
