--- tests/ros1/__init__.py
@@ -1,11 +1,0 @@
-import unittest
-
-from .test_interface import TestRos1Interface
-
-__all__ = [
-    "TestRos1Interface",
-]
-
-
-if __name__ == "__main__":
-    unittest.main()
