--- tests/ros1/node/dummy_node2.py
@@ -1,12 +1,0 @@
-import rospy
-from std_msgs.msg import String
-
-
-def main():
-    rospy.init_node("dummy_node2", disable_rosout=True)
-    rospy.Subscriber("/topic", String)
-    rospy.spin()
-
-
-if __name__ == "__main__":
-    main()
