+++ tests/view/test_info_panel.py
@@ -4,9 +4,9 @@ from textual.app import App, ComposeResult
 from textual.containers import ScrollableContainer
 from textual.widgets import Static
 
-from rtui_app.event import RosEntitySelected
-from rtui_app.ros import RosClient, RosEntity, RosEntityType
-from rtui_app.widgets import RosEntityInfoPanel
+from rtui2.event import RosEntitySelected
+from rtui2.ros import RosClient, RosEntity, RosEntityType
+from rtui2.widgets import RosEntityInfoPanel
 
 
 class ExampleApp(App):
